+++ frontend/src/redux/store.ts
@@ -5,13 +5,15 @@ import { rootSaga } from "./rootSaga";
 import { profileReducer } from "./ducks/profile/profile.slice";
 import { postReducer } from "./ducks/post/post.slice";
 import { userReducer } from "./ducks/user/user.slice";
+import { roleReducer } from "./ducks/role/role.slice";
 
 
 const rootReducer = combineReducers({
 	authReducer,
 	profileReducer,
 	postReducer,
-	userReducer
+	userReducer,
+	roleReducer
 });
 
 const sagaMiddleware = createSagaMiddleware();